+++ tests/test_erosion/test_mesh_processor.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from dfastbe.bank_erosion.data_models.mesh import MeshData, RiverSegment
+from dfastbe.bank_erosion.mesh import MeshData, RiverSegment
 from dfastbe.bank_erosion.mesh_processor import (
     MeshProcessor,
     enlarge,
